Refactor: UI 코드 분리
- 도메인별 UI 분리
