+++ src/views/common/elements.js
@@ -1,0 +1,6 @@
+export const renderMainTitle = () => {
+  const titleElement = document.createElement('h1');
+  titleElement.textContent = '🎱 행운의 로또';
+
+  return titleElement;
+};
